% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillCustomLinkagePlan.R
\name{fillCustomLinkagePlan}
\alias{fillCustomLinkagePlan}
\title{Fill matrix of true.items with values from a custom linkage plan}
\usage{
fillCustomLinkagePlan(t.tot, true.items, linkage.plan, min.a, max.a, mu.b,
  sd.b, ...)
}
\arguments{
\item{t.tot}{number of tests}

\item{true.items}{matrix containing all items in the rows and all test forms
in the columns}

\item{linkage.plan}{matrix containing the number of items each test form has
in common with itself and another test form}

\item{min.a}{Lower bound for the (uniform) distribution of item
discrimination parameter}

\item{max.a}{Upper bound for the (uniform) distribution of item
discrimination parameter}

\item{mu.b}{Mean of (normally-distributed) item difficulty parameter}

\item{sd.b}{Standard deviation of (normally-distributed) item difficulty
parameter}

\item{...}{Arguments to pass to genItemParameter}
}
\description{
Fill matrix of true.items with values from a custom linkage plan
}
