% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillRandomLinkagePlan.R
\name{fillRandomLinkagePlan}
\alias{fillRandomLinkagePlan}
\title{Fill matrix of true.items with values for a random linkage plan}
\usage{
fillRandomLinkagePlan(t.tot, J, C, true.items, min.a, max.a, mu.b, sd.b,
  ...)
}
\arguments{
\item{t.tot}{number of tests}

\item{J}{number of total items (unique + common) per test}

\item{C}{number of common items between two test}

\item{true.items}{matrix containing all items in the rows and all test forms
in the columns}

\item{min.a}{Lower bound for the (uniform) distribution of item
discrimination parameter}

\item{max.a}{Upper bound for the (uniform) distribution of item
discrimination parameter}

\item{mu.b}{Mean of (normally-distributed) item difficulty parameter}

\item{sd.b}{Standard deviation of (normally-distributed) item difficulty
parameter}

\item{...}{Arguments to pass to genItemParameter}
}
\description{
Fill matrix of true.items with values for a random linkage plan
}
