% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genAbilityParms.R
\name{genAbilityParms}
\alias{genAbilityParms}
\title{Generate True Ability Parameters}
\usage{
genAbilityParms(forms.n, mu.first, sd.first, mu.last = mu.first,
  sd.last = sd.first)
}
\arguments{
\item{forms.n}{number of forms}

\item{mu.first}{mean ability of first form}

\item{sd.first}{standard deviation of ability of first form}

\item{mu.last}{mean ability of last form}

\item{sd.last}{standard deviation of ability of last form}
}
\value{
List containing means and standard deviations per form
}
\description{
Generate True Ability Parameters
}
