% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genExamAnswers.R
\name{genExamAnswers}
\alias{genExamAnswers}
\title{Simulate exam test answers}
\usage{
genExamAnswers(true.item.parms, true.skills, join.anchors = FALSE)
}
\arguments{
\item{true.item.parms}{true item parameters per form}

\item{true.skills}{true examinee skill per form}

\item{join.anchors}{join answers for anchor items on each test?}
}
\value{
List of size equal to the number of forms. Each element of the list
  contains a matrix of dichotomic answers (right or wrong) per examinee and
  item.
}
\description{
Simulates answers for a test, given the true item parameters
  (2PL model) as well as the examinee skills per test.
}
