% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genItemBankExt.R
\name{genItemBankExt}
\alias{genItemBankExt}
\title{Generate item bank for NEAT with external anchor items}
\usage{
genItemBankExt(C, U, t.tot, min.a, max.a, mu.b, sd.b)
}
\arguments{
\item{C}{number of common items between two tests}

\item{U}{number of unique items per test}

\item{t.tot}{number of forms}

\item{min.a}{Lower bound for the (uniform) distribution of item
discrimination parameter}

\item{max.a}{Upper bound for the (uniform) distribution of item
discrimination parameter}

\item{mu.b}{Mean of (normally-distributed) item difficulty parameter}

\item{sd.b}{Standard deviation of (normally-distributed) item difficulty
parameter}
}
\value{
list containing 2PL item parameters per form
}
\description{
Generate item bank for NEAT with external anchor items
}
