% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genItemParameter.R
\name{genItemParameter}
\alias{genItemParameter}
\title{Generate Item Parameter Values}
\usage{
genItemParameter(type, parms, quantity = 1, truncate.b = c(-Inf, Inf))
}
\arguments{
\item{type}{Either "a" for the item discrimination or "b" for the item
difficulty}

\item{parms}{vector of length two. If type = "a", parms must be setup such
that \eqn{a ~ U(parms[1], parms[2])}. For type = "b", parms must be such
that \eqn{b ~ N(parms[1], parms[2])}.}

\item{quantity}{number of item parameters to generate}

\item{truncate.b}{vector of length 2 containing the limits for the difficulty
distribution}
}
\value{
scalar containing one item parameter
}
\description{
Generate Item Parameter Values
}
