% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTrueSkills.R
\name{genTrueSkills}
\alias{genTrueSkills}
\title{Simulates true examinee skills per form}
\usage{
genTrueSkills(I, num.forms, mu.skill = rep(0, num.forms),
  sd.skill = rep(1, num.forms))
}
\arguments{
\item{I}{number of test takers per form. It can have length 1, so all forms
will have the same number of test takers, or be a larger vector with the
respective number of examinees on each form}

\item{num.forms}{number of test forms}

\item{mu.skill}{mean skill per form}

\item{sd.skill}{standard deviation of skills per form}
}
\value{
List of examinee skill per form
}
\description{
Generates a matrix containing the true (latent) skill per
  administration for a given group of examinees.
}
