% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probIRT.R
\name{probIRT}
\alias{probIRT}
\title{Probability of Getting an Item Correct}
\usage{
probIRT(theta, a, b)
}
\arguments{
\item{theta}{examinee skill}

\item{a}{item discimination}

\item{b}{item difficulty}
}
\value{
Probability of getting a correct item.
}
\description{
Calculates the probability of getting an item correct under a
  2-parameter IRT model.
}
